create executeCommand function
